+++ common/src/main/java/ink/astrius/create_avionics/compat/simulated/peripherals/NavTablePeripheral.java
@@ -53,7 +53,7 @@ public class NavTablePeripheral extends SimPeripheral<NavTableBlockEntity> {
     }
 
     @LuaFunction
-    public Float getRelativeAngle() {
+    public double getRelativeAngle() {
         return this.blockEntity.getRelativeAngle();
     }
